Revamped createBMP to reduce the number of scripts. Fixes #13

+++ createBMP.sh
@@ -18,34 +18,53 @@ readonly vramsize=10
 readonly diskSizeInGiB=15
 
 # Delete it if its already there
-vboxmanage list runningvms | grep -q "\"${VM:?}\"" && {
+if $(vboxmanage list runningvms | grep --quiet "\"${VM:?}\"")
+then
     vboxmanage controlvm ${VM:?} poweroff
-}
-vboxmanage list vms | grep -q "\"${VM:?}\"" && {
-    vboxmanage unregistervm ${VM:?} --delete
-}
+    rebootBMP
+else
+    createBMP
+fi
 
 # Create the VM
-vboxmanage createvm --name "${VM:?}" -ostype "${os:?}"  --register
-# Attach a sata storage controller
-vboxmanage storagectl "${VM:?}"  --name "${storage:?}" --add sata
+function createBMP() {
+    vboxmanage createvm --name "${VM:?}" --ostype "${os:?}"  --register
+    # Attach a sata storage controller
+    vboxmanage storagectl "${VM:?}"  --name "${storage:?}" --add sata
 
-# Create a hard drive and attach it to the vm
-vboxmanage createhd --filename "${diskfile:?}" --size $((diskSizeInGiB * 1024))
-vboxmanage storageattach "${VM}" --storagectl "${storage:?}" --port 0 --device 0 --type hdd --medium "${diskfile:?}"
+    # Create a hard drive and attach it to the vm
+    vboxmanage createhd --filename "${diskfile:?}" --size $((diskSizeInGiB * 1024))
+    vboxmanage storageattach "${VM}" --storagectl "${storage:?}" --port 0 --device 0 --type hdd --medium "${diskfile:?}"
 
-# Modify the vm to boot from disk, and then from the network
-vboxmanage modifyvm "${VM:?}" --boot1 disk --boot2 net --boot3 none --boot4 none
-# Modify the vm to use an adapter that supports PXE booting
-VBoxManage modifyvm "${VM:?}" --nictype1 Am79C973
-# Connect the VM to the hostonly network that we're using
-vboxmanage modifyvm "${VM:?}" --nic1 hostonly --hostonlyadapter1 $(get_private_network_name)
-# Modify the vms' memory and vram sizes
-vboxmanage modifyvm "${VM:?}" --memory "${memsize:?}" --vram "${vramsize:?}"
-# Report the vm's MAC address
-MAC=$(vboxmanage showvminfo "${VM:?}" | grep 'NIC 1' |
-sed -n "/MAC/s/.*MAC: \([^,][^,]*\),.*/\1/p" |
-sed "s/\(..\)\(..\)\(..\)\(..\)\(..\)\(..\)/\1:\2:\3:\4:\5:\6/")
-echo $MAC,192.168.50.196,bmp.lab.net >.dhcp/dhcp-hostsfile
-vagrant ssh dhcp -c 'sudo service dnsmasq restart'
-echo "${VM:?}'s MAC address is: ${MAC:?}"
+    # Modify the vm to boot from disk, and then from the network
+    vboxmanage modifyvm "${VM:?}" --boot1 disk --boot2 net --boot3 none --boot4 none
+    # Modify the vm to use an adapter that supports PXE booting
+    VBoxManage modifyvm "${VM:?}" --nictype1 Am79C973
+    # Connect the VM to the hostonly network that we're using
+    vboxmanage modifyvm "${VM:?}" --nic1 hostonly --hostonlyadapter1 $(get_private_network_name)
+    # Modify the vms' memory and vram sizes
+    vboxmanage modifyvm "${VM:?}" --memory "${memsize:?}" --vram "${vramsize:?}"
+    # Report the vm's MAC address
+    MAC=$(vboxmanage showvminfo "${VM:?}" | grep 'NIC 1' |
+		 sed -n "/MAC/s/.*MAC: \([^,][^,]*\),.*/\1/p" |
+		 sed "s/\(..\)\(..\)\(..\)\(..\)\(..\)\(..\)/\1:\2:\3:\4:\5:\6/")
+    echo $MAC,${IP:?},${HOSTNAME:?} >.dhcp/dhcp-hostsfile
+    vagrant ssh dhcp -c 'sudo service dnsmasq restart'
+    echo "${VM:?}'s MAC address is: ${MAC:?}"
+}
+
+function rebootBMP() {
+    # Detach and delete the disk
+    vboxmanage storageattach "${VM}" --storagectl "${storage:?}" --port 0 --device 0 --type hdd --medium none
+    vboxmanage closemedium disk ${diskfile:?} --delete
+
+    # Create a new blank disk and attach it
+    vboxmanage createhd --filename "${diskfile:?}" --size $((diskSizeInGiB * 1024))
+    vboxmanage storageattach "${VM}" --storagectl "${storage:?}" --port 0 --device 0 --type hdd --medium "${diskfile:?}"
+
+    # Delete the agent from the OMS
+    vagrant ssh oms -c "sudo su -c \"/vagrant/deleteAgent.sh ${HOSTNAME:?}\" - oracle"
+
+    # start the vm
+    vboxmanage startvm ${VM}
+}
